FDAM_mmseg/requirements.txt
absl-py==1.0.0
addict==2.4.0
albumentations==1.1.0
appdirs==1.4.4
cachetools==5.0.0
charset-normalizer==2.0.12
cityscapesScripts==2.2.1
click==8.1.8
coloredlogs==15.0.1
cycler==0.11.0
Cython==0.29.28
efficientnet-pytorch==0.6.3
einops==0.6.1
ever-beta==0.2.3
flash-attn==0.2.8
fonttools==4.30.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
humanfriendly==10.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.0
Markdown==3.3.6
markdown-it-py==2.2.0
mat4py==0.6.0
matplotlib==3.5.1
mdurl==0.1.2
mmcls==0.23.1
mmcv-full==1.5.3
mmengine==0.10.7
mmpretrain==1.2.0
mmsegmentation==0.25.0
model-index==0.1.11
modelindex==0.0.2
munch==2.5.0
networkx==2.6.3
ninja==1.11.1.3
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
ordered-set==4.1.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pretrainedmodels==0.7.4
prettytable==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
Pygments==2.17.2
pyparsing==3.0.7
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.8.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.2.1
SimpleCV==1.3
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.11.0+cu113
torch-dct==0.1.6
torchaudio==0.7.2
torchprofile==0.0.4
torchvision==0.12.0+cu113
tqdm==4.63.0
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
yapf==0.32.0
zipp==3.7.0
